% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_class_viz.R
\name{clust_class_viz}
\alias{clust_class_viz}
\title{Title visual contingency table of the true classification (partitioning) and the clustering results}
\usage{
clust_class_viz(object, features = NULL, clusters = NULL)
}
\arguments{
\item{object}{: Clustering result from usual methods of clustering}

\item{features}{: True Partition (Classification) of the observation}

\item{clusters}{: Possibility to display a visual contingency table from just clustering results}
}
\value{
a GGalluvial graph
}
\description{
Title visual contingency table of the true classification (partitioning) and the clustering results
}
\examples{
:
# depencies :
#library("ggplot2")
#library("ggalluvial")
# features : vector of factors
# clusters : vector of factors
# Data preparation
# +++++++++++++++
# data("iris")
# head(iris)
# # Remove species column (5) and scale the data
iris.scaled <- scale(iris[, -5])
# true_class = iris[,5]
# # methods from the package stats
# library(stats)
# # (kmeans method)
# km.res <- kmeans(iris.scaled, 5, nstart = 10)
# clust_class_viz(km.res, true_class)
# # ------------------------------------------------------------------------------
# # methods from the package cluster
# library(cluster)
# # (pam method)
# pam.res = pam(iris.scaled, 3)
# clust_class_viz(pam.res, true_class)
# # (clara method)
# clara.res = clara(iris.scaled, 3)
# clust_class_viz(clara.res, true_class)
# # (fanny method)
# fanny.res = fanny(iris.scaled, 3)
# clust_class_viz(fanny.res, true_class)
# # ------------------------------------------------------------------------------
# # methods from the package fpc
# library(fpc)
# # (dbscan method)
# dbscan.res = dbscan(iris.scaled, eps = 0.2)
# clust_class_viz(dbscan.res, true_class)
# # dbscan method don't provide information, clusters are
# # required for external measures computation
# # ------------------------------------------------------------------------------
# # methods from the package mclust
# library(mclust)
# # Mclust method
# Mclust.res = Mclust(iris.scaled,4)
# clust_class_viz(Mclust.res, true_class)
# # ------------------------------------------------------------------------------
# # methods from the package FactoMineR
# library(FactoMineR)
# # HCPC method
# HCPC.res = HCPC(data.frame(iris.scaled),3)
# clust_class_viz(HCPC.res,true_class)
# # ------------------------------------------------------------------------------
# # methods from the package factoextra
# library(factoextra)
# # hkmeans method
# hkmeans.res = hkmeans(iris.scaled,3)
# clust_class_viz(hkmeans.res,true_class)
# ------------------------------------------------------------------------------
# using the two arguments of the function : features, clusters
#clust_class_viz(object = NULL,features = true_class, clusters=HCPC.res$call$X$clust)
}
\author{
Mohamed Al Mahdi Tantaoui
}
