% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_radar_viz.R
\name{clust_radar_viz}
\alias{clust_radar_viz}
\title{Title Variables means by clusters}
\usage{
clust_radar_viz(object, data = NULL, clusters = NULL)
}
\arguments{
\item{object}{: Clustering result from usual methods of clustering}

\item{data}{: the data which was used to compute the clustering methods results}

\item{clusters}{Possibility to display a radar of the clustering from just clustering results and the original data}
}
\value{
a radar plot
}
\description{
Title Variables means by clusters
}
\examples{
:

# depencies :
# devtools::install_github("ricardo-bion/ggradar", dependencies = TRUE)
# library(devtools)
# library(digest)
# library(ggradar)
# library(scales)
# library(dplyr)
# library(purrrlyr)
#
# data("iris")
# head(iris)
# # Remove species column (5) and scale the data
# iris.scaled <- scale(iris[, -5])
# true_class = iris[,5]
# # methods from the package stats
# library(stats)
# (kmeans method)
# km.res <- kmeans(iris.scaled, 3, nstart = 10)
# #clust_radar_viz(km.res, data = iris[,1:4])
# # ------------------------------------------------------------------------------
# # methods from the package cluster
# library(cluster)
# # (pam method)
# pam.res = pam(iris.scaled, 5)
# clust_radar_viz(pam.res)
# # (clara method)
# clara.res = clara(iris.scaled, 3)
# #clust_radar_viz(clara.res)
# # (fanny method)
# fanny.res = fanny(iris.scaled, 3)
# #clust_radar_viz(fanny.res)
# # ------------------------------------------------------------------------------
# # methods from the package fpc
# library(fpc)
# # (dbscan method)
# dbscan.res = dbscan(iris.scaled, eps = .4, MinPts = 4)
# #clust_radar_viz(object = dbscan.res, data = iris[,1:4])
# # dbscan method don't provide information, clusters are
# # required for external measures computation
# # ------------------------------------------------------------------------------
# # methods from the package mclust
# library(mclust)
# # Mclust method
# Mclust.res = Mclust(iris.scaled,4)
# #clust_radar_viz(Mclust.res)
# # ------------------------------------------------------------------------------
# # methods from the package FactoMineR
# library(FactoMineR)
# # HCPC method
# HCPC.res = HCPC(data.frame(iris.scaled),3)
# #clust_radar_viz(HCPC.res)
# # ------------------------------------------------------------------------------
# # methods from the package factoextra
# library(factoextra)
#
# # hkmeans method
# hkmeans.res = hkmeans(iris.scaled,3)
# #clust_radar_viz(hkmeans.res)
# # ------------------------------------------------------------------------------
# # using the two arguments of the function : labels, clusters
# # clust_radar_viz(object = NULL, data = iris.scaled, clusters=km.res$cluster)
}
\author{
Mohamed Al Mahdi Tantaoui
}
