% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_summary_viz.R
\name{clust_summary_viz}
\alias{clust_summary_viz}
\title{Title : Entropy based Measures (external measures)}
\usage{
clust_summary_viz(object, data = NULL, clusters = NULL)
}
\arguments{
\item{object}{: Clustering result from usual methods of clustering}

\item{data}{: the data which was used to compute the clustering methods results}

\item{clusters}{: Possibility to display a summary of the clustering from just clustering results and the original data}
}
\value{
Clustering Summary
}
\description{
Title : Entropy based Measures (external measures)
}
\examples{
:
# # depencies
# library(mlr)
# library(tidyverse)
# library(ggplot2)
# library(GGally)
# library(plotly)
# library(dplyr)
# library(factoextra)
# library(stats)
#
# data("iris")
# head(iris)
# # Remove species column (5) and scale the data
# iris.scaled <- scale(iris[, -5])
# true_class = iris[,5]
# # methods from the package stats
# library(stats)
# # (kmeans method)
# km.res <- kmeans(iris.scaled, 3, nstart = 10)
# clust_summary_viz(km.res, iris.scaled)
# # ------------------------------------------------------------------------------
# # methods from the package cluster
# library(cluster)
# # (pam method)
# pam.res = pam(iris.scaled, 5)
# clust_summary_viz(pam.res)
# # (clara method)
# clara.res = clara(iris.scaled, 3)
# clust_summary_viz(clara.res)
# # (fanny method)
# fanny.res = fanny(iris.scaled, 3)
# clust_summary_viz(fanny.res)
# # ------------------------------------------------------------------------------
# # methods from the package fpc
# library(fpc)
# # (dbscan method)
# dbscan.res = dbscan(iris.scaled, eps = .4, MinPts = 4)
# clust_summary_viz(object = dbscan.res, data = iris[,1:4])
# # dbscan method don't provide information, clusters are
# # required for external measures computation
# # ------------------------------------------------------------------------------
# # methods from the package mclust
# library(mclust)
# # Mclust method
# Mclust.res = Mclust(iris.scaled,4)
# clust_summary_viz(Mclust.res)
# # ------------------------------------------------------------------------------
# # methods from the package FactoMineR
# library(FactoMineR)
# # HCPC method
# HCPC.res = HCPC(data.frame(iris.scaled),3)
# clust_summary_viz(HCPC.res)
# # ------------------------------------------------------------------------------
# # methods from the package factoextra
# library(factoextra)
# # hkmeans method
# hkmeans.res = hkmeans(iris.scaled,3)
# clust_summary_viz(hkmeans.res)
# # ------------------------------------------------------------------------------
# # using the two arguments of the function : labels, clusters
# clust_summary_viz(object = NULL, data = iris.scaled, clusters=dbscan.res$cluster)
}
\author{
Mohamed Al Mahdi Tantaoui
}
