% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_viz_3d.R
\name{clust_viz_3d}
\alias{clust_viz_3d}
\title{Title 3D reprenstation of the Clusters}
\usage{
clust_viz_3d(object, data = NULL, clusters = NULL, stand = TRUE)
}
\arguments{
\item{object}{: Clustering result from usual methods of clustering}

\item{data}{: the data which was used to compute the clustering methods results}

\item{clusters}{: : Possibility to display in 3D the clustering  from just the vector of clustering results}

\item{stand}{: Boolean used to skip the scaling procedure for the HCPC method results}
}
\value{
3D representation of the clustering results
}
\description{
Title 3D reprenstation of the Clusters
}
\examples{
:

# depencies :
# library(ggplot2)
# library(plotly)
# library(dplyr)
# library(factoextra)
# library(stats)
#
# data("iris")
# head(iris)
# # Remove species column (5) and scale the data
# iris.scaled <- scale(iris[, -5])
# true_class = iris[,5]
# # methods from the package stats
# library(stats)
# # (kmeans method)
# km.res <- kmeans(iris.scaled, 3, nstart = 10)
# clust_viz_3d(km.res, data = iris.scaled)
# # ------------------------------------------------------------------------------
# # methods from the package cluster
# library(cluster)
# # (pam method)
# pam.res = pam(iris.scaled, 5)
# clust_viz_3d(pam.res)
# # (clara method)
# clara.res = clara(iris.scaled, 3)
# clust_viz_3d(clara.res)
# # (fanny method)
# fanny.res = fanny(iris.scaled, 3)
# clust_viz_3d(fanny.res)
# # ------------------------------------------------------------------------------
# # methods from the package fpc
# library(fpc)
# # (dbscan method)
# dbscan.res = dbscan(iris.scaled, eps = .4, MinPts = 4)
# clust_viz_3d(object = dbscan.res, data = iris.scaled)
# # dbscan method don't provide information, clusters are
# # required for external measures computation
# # ------------------------------------------------------------------------------
# # methods from the package mclust
# library(mclust)
# # Mclust method
# Mclust.res = Mclust(iris.scaled,4)
# clust_viz_3d(Mclust.res)
# # ------------------------------------------------------------------------------
# # methods from the package FactoMineR
# library(FactoMineR)
# # HCPC method
# HCPC.res = HCPC(data.frame(iris.scaled),3)
# clust_viz_3d(HCPC.res)
# # ------------------------------------------------------------------------------
# # methods from the package factoextra
# library(factoextra)
# # hkmeans method
# hkmeans.res = hkmeans(iris.scaled,3)
# clust_viz_3d(hkmeans.res)
# # ------------------------------------------------------------------------------
# # using the two arguments of the function : labels, clusters
# clust_viz_3d(object = NULL, data = iris.scaled, clusters=km)
}
\author{
Mohamed Al Mahdi Tantaoui
}
