% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Khi2.R
\name{Khi2}
\alias{Khi2}
\title{Title Khi2}
\usage{
Khi2(df, col)
}
\arguments{
\item{df}{: First qualitative variable}

\item{col}{: Second qualitative variable}
}
\value{
the Kramer's V
}
\description{
Title Khi2
}
\examples{
# # depencies :
# library(cowplot)
# library(ggplot2)
####### Package
#--install.packages("ggplot2")
#install.packages("cowplot")
#######Library
# library(ggplot2)
# library(cowplot)
######## Recuperer les données et kmeans ######
# data(iris)
# dat <- subset(iris, Species != "setosa")
# spe <- iris[,5]
# res.km <- kmeans(scale(iris[, -5]), 3, nstart = 25)
# clus<-res.km$cluster
############################# Khi 2 - V Cramer  ############################
}
\author{
Mohamed Al Mahdi Tantaoui
}
