% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate.MoyCon.R
\name{stat.comp}
\alias{stat.comp}
\title{Title Conditional mean of the variables to the classes}
\usage{
stat.comp(x, y)
}
\arguments{
\item{obj}{: equals to another function "Univarie.MoyCon(data,cluster)" with the data at first, and then classes in second}

\item{seuil}{: Treshold}
}
\value{
Conditional mean
}
\description{
Title Conditional mean of the variables to the classes
}
\examples{
# # depencies :
# library(cowplot)
# library(ggplot2)
####### Package
#--install.packages("ggplot2")
#install.packages("cowplot")
#######Library
# library(ggplot2)
# library(cowplot)
######## Recuperer les données et kmeans ######
# data(iris)
# dat <- subset(iris, Species != "setosa")
# res.km <- kmeans(scale(iris[, -5]), 3, nstart = 25)
# clus<-res.km$cluster
####### Moy Con ########
# obj <- Univarie.MoyCon(iris[,-5],res.km$cluster)
# print(class(obj))
# print(obj)
# graphique.MoyCon(obj,seuil=70)
############################ MOYENNE CONDITIONELLE##################################
#Recuperer la moyenne condition pour savoir l'importance de chaque variable
}
\author{
Mohamed Al Mahdi Tantaoui
}
