% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestQuanti.R
\name{TestQuanti}
\alias{TestQuanti}
\title{Title Test between quantitative and qualitative variables}
\usage{
TestQuanti(df, col)
}
\arguments{
\item{df}{: quantitative(s) variable(s)}

\item{col}{: qualitative variable}
}
\value{
ANOVA or Student test
}
\description{
Title Test between quantitative and qualitative variables
}
\examples{
# # depencies :
# library(cowplot)
# library(ggplot2)
####### Package
#--install.packages("ggplot2")
#install.packages("cowplot")
#######Library
#library(ggplot2)
#library(cowplot)
######## Recuperer les données et kmeans ######
# data(iris)
# dat <- subset(iris, Species != "setosa")
# spe <- iris[,5]
# res.km <- kmeans(scale(iris[, -5]), 3, nstart = 25)
# clus<-res.km$cluster
###### Test ########
#print(TestQuanti(iris[,-5],res.km$cluster))
#TestQuanti(iris[,-5],res.km$cluster)
######################## Student vs ANOVA #################################
}
\author{
Mohamed Al Mahdi Tantaoui
}
